简单跑通
解冻是否只有撤单引起？

+++ bolt-spot/src/main/java/com/cmex/bolt/spot/domain/Currency.java
@@ -6,6 +6,6 @@ import lombok.Data;
 @Data
 @Builder
 public class Currency {
-    private short id;
+    private int id;
     private String name;
 }
