+++ bolt-spot/src/main/java/com/cmex/bolt/spot/repository/impl/OrderBookRepository.java
@@ -2,5 +2,5 @@ package com.cmex.bolt.spot.repository.impl;
 
 import com.cmex.bolt.spot.domain.OrderBook;
 
-public class OrderBookRepository extends HashMapRepository<Short, OrderBook>{
+public class OrderBookRepository extends HashMapRepository<Integer, OrderBook>{
 }
